requirements.txt
sqlalchemy
pymysql
tabulate
fpdf2
pandas
openpyxl

prompt_toolkit
